+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31" >
         <activity
-            android:name=".ui.library.act.Library"
+            android:name=".ui.library.act.LibraryActivity"
             android:exported="false"
             android:label="@string/title_activity_library"
             android:theme="@style/Theme.PlaylistMaker.NoActionBar" />
