+++ app/src/main/java/com/example/playlistmaker/ui/main/act/MainActivity.kt
@@ -9,8 +9,7 @@ import androidx.appcompat.app.AppCompatDelegate
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 import com.example.playlistmaker.databinding.ActivityMainBinding
-import com.example.playlistmaker.ui.library.act.Library
-import com.example.playlistmaker.ui.player.act.MediaActivity
+import com.example.playlistmaker.ui.library.act.LibraryActivity
 import com.example.playlistmaker.ui.search.act.SearchActivity
 import com.example.playlistmaker.ui.settings.act.SettingsActivity
 import com.example.playlistmaker.ui.main.vm.MainActivityViewModel
@@ -49,7 +48,7 @@ class MainActivity : AppCompatActivity() {
 
         binding.medialibraryButton.setOnClickListener {
 
-            val mediaIntent = Intent(this, Library::class.java)
+            val mediaIntent = Intent(this, LibraryActivity::class.java)
             startActivity(mediaIntent)
         }
