+++ app/src/main/res/layout/activity_library.xml
@@ -1,66 +1,64 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-
     android:background="@color/settings_screen_background"
     android:orientation="vertical">
 
+    <!--
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/media_library_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="56dp"
+        android:navigationIcon="@drawable/back_icon"
+        style="@style/main_font"
+        android:title="@string/media_library_header">
+    </androidx.appcompat.widget.Toolbar>
+    -->
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/settings_back_button_click_area">
+
+        <!-- BIGGER CLICK AREA -->
         <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/settings_back_button_click_area">
+            android:id="@+id/media_library_back_button"
+            android:layout_width="@dimen/settings_back_button_click_area"
+            android:layout_height="@dimen/settings_back_button_click_area" />
 
-                <!-- BIGGER CLICK AREA -->
-                <FrameLayout
-                    android:id="@+id/media_library_back_button"
-                    android:layout_width="@dimen/settings_back_button_click_area"
-                    android:layout_height="@dimen/settings_back_button_click_area"
-                />
+        <ImageView
+            android:layout_width="16dp"
+            android:layout_height="16dp"
+            android:layout_marginStart="20dp"
+            android:layout_marginTop="20dp"
+            android:contentDescription="@string/settings_back_button_description"
+            android:src="@drawable/back_icon" />
 
-                <ImageView
-                    android:layout_width="16dp"
-                    android:layout_height="16dp"
+        <TextView
+            style="@style/main_font"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="64dp"
+            android:layout_marginTop="14dp"
+            android:text="@string/media_library_header"
+            android:textColor="@color/settings_screen_header" />
+    </FrameLayout>
 
-                    android:layout_marginStart="20dp"
-                    android:layout_marginTop="20dp"
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/media_library_tabs"
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:background="@color/settings_screen_background"
+        app:tabIndicatorColor="@color/settings_screen_header"
+        app:tabRippleColor="@color/main_screen_background"
+        app:tabTextAppearance="@style/media_player_time_code_font"
+        app:tabTextColor="@color/settings_screen_header" />
 
-                    android:src="@drawable/back_icon"
-                    android:contentDescription="@string/settings_back_button_description"
-                />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-
-                    android:layout_marginStart="64dp"
-                    android:layout_marginTop="14dp"
-
-                    android:text="@string/media_library_header"
-                    android:textColor="@color/settings_screen_header"
-                    style="@style/main_font"
-                />
-        </FrameLayout>
-
-
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/media_library_tabs"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-
-            android:background="@color/settings_screen_background"
-            app:tabTextColor="@color/settings_screen_header"
-            app:tabIndicatorColor="@color/settings_screen_header"
-        />
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/media_library_view_pager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-        />
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/media_library_view_pager"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
 </LinearLayout>
