+++ app/src/main/res/layout/fragment_fav_songs.xml
@@ -1,35 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-        <FrameLayout
-            android:layout_width ="wrap_content"
+    <FrameLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center">
+
+        <ImageView
+            android:layout_width="120dp"
+            android:layout_height="120dp"
+            android:layout_gravity="center_horizontal"
+            android:contentDescription="@string/search_info_no_data"
+            android:src="@drawable/no_data" />
+
+        <TextView
+            android:id="@+id/fav_song_frag_text"
+            style="@style/search_info_text"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-        >
-                <ImageView
-                    android:layout_width ="120dp"
-                    android:layout_height="120dp"
-                    android:layout_gravity="center_horizontal"
-                    android:src="@drawable/no_data"
-                />
-
-                <TextView
-                    android:id="@+id/fav_song_frag_text"
-
-                    android:layout_width ="wrap_content"
-                    android:layout_height="wrap_content"
-
-                    android:layout_gravity="center_horizontal"
-                    android:layout_marginTop="128dp"
-
-                    android:text="**********"
-                    style="@style/search_info_text"
-                    android:textColor="@color/search_screen_header"
-                />
-        </FrameLayout>
+            android:layout_gravity="center_horizontal"
+            android:layout_marginTop="128dp"
+            android:text=""
+            android:textColor="@color/search_screen_header" />
+    </FrameLayout>
 
 </FrameLayout>
